attempt to fix tests that fail due to JEP-181
