+++ src/test/java/com/hivemq/extensions/handler/ClientLifecycleEventHandlerTest.java
@@ -394,8 +394,8 @@ public class ClientLifecycleEventHandlerTest {
     private ClientLifecycleEventListenerProvider getTestProvider(CountDownLatch countDownLatch) throws Exception {
 
         final JavaArchive javaArchive = ShrinkWrap.create(JavaArchive.class)
-                .addClass("com.hivemq.extensions.handler.ClientLifecycleEventHandlerTest$TestProvider")
-                .addClass("com.hivemq.extensions.handler.ClientLifecycleEventHandlerTest$TestProvider$1");
+                .addClass("com.hivemq.extensions.handler.testextensions.TestProvider")
+                .addClass("com.hivemq.extensions.handler.testextensions.TestProvider$1");
 
         final File jarFile = temporaryFolder.newFile();
         javaArchive.as(ZipExporter.class).exportTo(jarFile, true);
@@ -403,78 +403,9 @@ public class ClientLifecycleEventHandlerTest {
         //This classloader contains the classes from the jar file
         final IsolatedPluginClassloader cl = new IsolatedPluginClassloader(new URL[]{jarFile.toURI().toURL()}, this.getClass().getClassLoader());
 
-        final Class<?> providerClass = cl.loadClass("com.hivemq.extensions.handler.ClientLifecycleEventHandlerTest$TestProvider");
+        final Class<?> providerClass = cl.loadClass("com.hivemq.extensions.handler.testextensions.TestProvider");
 
         ClientLifecycleEventListenerProvider testProvider = (ClientLifecycleEventListenerProvider) providerClass.getDeclaredConstructor(CountDownLatch.class).newInstance(countDownLatch);
         return testProvider;
-    }
-
-    public static class TestProvider implements ClientLifecycleEventListenerProvider {
-
-        private final CountDownLatch onMqttConnectionStartLatch;
-        private final CountDownLatch onAuthenticationFailedDisconnectLatch;
-        private final CountDownLatch onConnectionLostLatch;
-        private final CountDownLatch onClientInitiatedDisconnectLatch;
-        private final CountDownLatch onServerInitiatedDisconnectLatch;
-        private final CountDownLatch onAuthenticationSuccessfulLatch;
-        private final CountDownLatch onDisconnectLatch;
-
-        public TestProvider(final CountDownLatch countDownLatch) {
-            this.onMqttConnectionStartLatch = countDownLatch;
-            this.onAuthenticationFailedDisconnectLatch = countDownLatch;
-            this.onConnectionLostLatch = countDownLatch;
-            this.onClientInitiatedDisconnectLatch = countDownLatch;
-            this.onServerInitiatedDisconnectLatch = countDownLatch;
-            this.onAuthenticationSuccessfulLatch = countDownLatch;
-            this.onDisconnectLatch = countDownLatch;
-        }
-
-        @Override
-        public @Nullable ClientLifecycleEventListener getClientLifecycleEventListener(@NotNull ClientLifecycleEventListenerProviderInput input) {
-
-            return new ClientLifecycleEventListener() {
-                @Override
-                public void onMqttConnectionStart(@NotNull ConnectionStartInput input) {
-                    onMqttConnectionStartLatch.countDown();
-                    System.out.println("connect");
-                }
-
-                @Override
-                public void onAuthenticationFailedDisconnect(@NotNull AuthenticationFailedInput input) {
-                    onAuthenticationFailedDisconnectLatch.countDown();
-                    System.out.println("auth failed");
-                }
-
-                @Override
-                public void onConnectionLost(@NotNull ConnectionLostInput input) {
-                    onConnectionLostLatch.countDown();
-                    System.out.println("connection lost");
-                }
-
-                @Override
-                public void onClientInitiatedDisconnect(@NotNull ClientInitiatedDisconnectInput input) {
-                    onClientInitiatedDisconnectLatch.countDown();
-                    System.out.println("client disconnect");
-                }
-
-                @Override
-                public void onServerInitiatedDisconnect(@NotNull ServerInitiatedDisconnectInput input) {
-                    onServerInitiatedDisconnectLatch.countDown();
-                    System.out.println("server disconnect");
-                }
-
-                @Override
-                public void onAuthenticationSuccessful(@NotNull AuthenticationSuccessfulInput input) {
-                    onAuthenticationSuccessfulLatch.countDown();
-                    System.out.println("auth success");
-                }
-
-                @Override
-                public void onDisconnect(@NotNull DisconnectEventInput input) {
-                    onDisconnectLatch.countDown();
-                    System.out.println("disconnect");
-                }
-            };
-        }
     }
 }
